Use `AffinePoint` where appropriate
