+++ ed448-goldilocks/src/sign.rs
@@ -89,6 +89,7 @@ pub use verifying_key::*;
 
 use crate::{CompressedEdwardsY, EdwardsPoint, EdwardsScalar};
 use elliptic_curve::array::Array;
+use elliptic_curve::group::GroupEncoding;
 
 /// Length of a secret key in bytes
 pub const SECRET_KEY_LENGTH: usize = 57;
@@ -118,7 +119,7 @@ impl From<InnerSignature> for Signature {
     fn from(inner: InnerSignature) -> Self {
         let mut s = [0u8; SECRET_KEY_LENGTH];
         s.copy_from_slice(&inner.s.to_bytes_rfc_8032());
-        Self::from_components(inner.r.compress(), s)
+        Self::from_components(inner.r.to_bytes(), s)
     }
 }
 
@@ -129,7 +130,10 @@ impl TryFrom<&Signature> for InnerSignature {
         let s_bytes: &Array<u8, _> = (signature.s_bytes()).into();
         let s = Option::from(EdwardsScalar::from_canonical_bytes(s_bytes))
             .ok_or(SigningError::InvalidSignatureSComponent)?;
-        let r = Option::from(CompressedEdwardsY::from(*signature.r_bytes()).decompress())
+        let r = CompressedEdwardsY::from(*signature.r_bytes())
+            .decompress()
+            .into_option()
+            .map(|point| point.to_edwards())
             .ok_or(SigningError::InvalidSignatureRComponent)?;
         Ok(Self { r, s })
     }
